Move abstract test back to `dependency-management` test fixtures
It depends on other fixtures in the `dependency-management` project.
